Move selectors into selecotrs file
