+++ src/App/AppContainer.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { createSelector } from 'reselect';
 import { fetchInitialData } from './../actions';
+import { backgroundSelector } from './../selectors';
 import { withRouter } from 'react-router-dom';
 import { push } from 'react-router-redux';
 import App from './App';
@@ -15,19 +15,6 @@ class AppContainer extends Component {
     return <App {...this.props} />;
   }
 }
-
-const orderSelector = state => state.queuedOrder;
-const backgroundSelector = createSelector(orderSelector, order => {
-  if (order === null) {
-    return null;
-  }
-
-  const product = order.order.products.find(
-    product => product.splash_image !== null
-  );
-
-  return product === undefined ? null : product.splash_image;
-});
 
 const mapStateToProps = state => ({
   title: state.title,
