+++ src/App/Recent/index.js
@@ -1,19 +1,7 @@
 import { connect } from 'react-redux';
-import { createSelector } from 'reselect';
 import { selectMember } from '../../actions';
+import { recentSelector } from './../../selectors';
 import Members from './../Members/Members';
-import { uniqBy, take } from 'lodash';
-
-const transactionsSelector = state => state.transactions;
-const recentSelector = createSelector(transactionsSelector, transactions =>
-  take(
-    uniqBy(
-      transactions.map(transaction => transaction.member),
-      member => member.id
-    ),
-    6 * 6
-  )
-);
 
 const mapStateToProps = state => ({
   members: recentSelector(state)
